+++ src/pages/Timeline/PostsItems/DeleteStyle.js
@@ -2,6 +2,8 @@ import styled from "styled-components";
 
 const CustomStyles = {
   content: {
+    // position: "fixed",
+    // zIndex: "1000",
     fontFamily: "Lato",
     width: "597px",
     height: "262px",
