+++ src/pages/Timeline/PostsItems/PostsList.js
@@ -1,4 +1,6 @@
 import { useNavigate } from "react-router-dom";
+import React, { useState, useRef, useEffect } from "react";
+import useAuth from "../../../hooks/userAuth";
 import {
   ReadContainer,
   ProfileContainer,
@@ -17,12 +19,15 @@ import {
   LinkUrl,
 } from "./SnippetStyle";
 import { FiEdit2 } from "react-icons/fi";
-import React, { useState, useRef, useEffect } from "react";
 import { FormInput } from "../TimelineStyles";
 import { FiTrash2 } from "react-icons/fi";
 import ReactHashtag from "@mdnm/react-hashtag";
 import api from "../../../services/api";
-import DeletePost from "./DeletePost";
+import Modal from "react-modal";
+import { ThreeDots } from "react-loader-spinner";
+import { Cancel, CustomStyles, Delete, Form } from "./DeleteStyle";
+
+Modal.setAppElement(".root");
 
 export default function PostsLists({
   likes,
@@ -35,15 +40,31 @@ export default function PostsLists({
   const [linkEdit, setLink] = useState();
   const [descEdit, setDesc] = useState();
   const [edit, setEdit] = useState(false);
-  const [deletePost, setDeletePost] = useState(false);
   const [loading, setLoading] = useState(false);
   const [modalIsOpen, setIsOpen] = useState(false);
   const ref = useRef();
   const navigation = useNavigate();
-  console.log(posts);
-  function handleDelete() {
-    setDeletePost(true);
+  const { auth } = useAuth();
+
+  function openModal() {
     setIsOpen(true);
+  }
+
+  function closeModal() {
+    setIsOpen(false);
+  }
+
+  async function handleDelete(id) {
+    document.location.reload(true);
+    setIsOpen(false);
+    setLoading(false);
+
+    try {
+      await api.deletePost(id, auth);
+      setLoading(false);
+    } catch (error) {
+      alert("Erro ao apagar o post. Tente novamente");
+    }
   }
 
   function handleClick(link) {
@@ -130,19 +151,35 @@ export default function PostsLists({
                 {user.id === post.userId ? (
                   <div>
                     <FiEdit2 onClick={(e) => editPost(e, post)} />
-                    <FiTrash2 onClick={(e) => handleDelete()} />
+                    <FiTrash2 onClick={openModal} />
+                    <Modal
+                      isOpen={modalIsOpen}
+                      onRequestClose={closeModal}
+                      style={CustomStyles}
+                    >
+                      <h2>
+                        Are you sure you want <br />
+                        to delete this post?{" "}
+                      </h2>
+                      <Form>
+                        <Cancel onClick={closeModal}>No, go back</Cancel>
+                        <Delete
+                          onClick={() => handleDelete(post.id)}
+                          disabled={loading}
+                        >
+                          {loading ? (
+                            <ThreeDots color="#ffffff" height={20} width={20} />
+                          ) : (
+                            "yes, delete it"
+                          )}
+                        </Delete>
+                      </Form>
+                    </Modal>
                   </div>
                 ) : (
                   <></>
                 )}
               </UsenameContainer>
-              {deletePost && (
-                <DeletePost
-                  postId={post.id}
-                  setIsOpen={setIsOpen}
-                  modalIsOpen={modalIsOpen}
-                />
-              )}
               {edit && postEditId === post.id ? (
                 <FormInput
                   value={descEdit}
